Move mapping code
